backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
bcrypt==4.1.1
cryptography==42.0.8

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2

# Blockchain
web3==6.11.3
eth-account==0.9.0

# External APIs
openai==1.3.7
google-cloud-speech==2.21.0
plaid-python==10.0.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Database
pgvector==0.2.4
